requirements.txt
# Core Dependencies
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.38.0

# Agent Development Kit (ADK)
# Note: Install ADK based on official documentation
# pip install git+https://github.com/google/adk-python.git

# PDF Processing
pypdf2>=3.0.0
pdfplumber>=0.10.0
pymupdf>=1.23.0

# API Integrations
arxiv>=2.0.0
requests>=2.31.0
httpx>=0.25.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# NLP & Text Processing
nltk>=3.8.0
spacy>=3.7.0
sentence-transformers>=2.2.0

# Knowledge Graph
networkx>=3.2.0
pyvis>=0.3.2

# Memory & Storage
redis>=5.0.0
chromadb>=0.4.0
faiss-cpu>=1.7.4

# Observability
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation>=0.41b0
prometheus-client>=0.19.0
structlog>=23.2.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0
pydantic>=2.5.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Utilities
rich>=13.7.0
tqdm>=4.66.0
tenacity>=8.2.0

# Web Framework (for API)
fastapi>=0.104.0
uvicorn>=0.24.0

# Async Support
aiohttp>=3.9.0
asyncio>=3.4.3

# Citation Management
bibtexparser>=1.4.0
scholarly>=1.7.0

# Jupyter (for demo)
jupyter>=1.0.0
ipykernel>=6.27.0

